restructure components directory to clarify usage and update all references in other files
