+++ frontend/app/events/page.tsx
@@ -1,8 +1,8 @@
 import { getCurrentSession } from "@/actions/users.server";
-import PageLayout from "@/components/PageLayout";
+import PageLayout from "@/components/shared/PageLayout";
 import { UserDisplayData } from "@/schemas/fe.schema";
 import { redirect } from "next/navigation";
-import EventsFrame from "@/components/EventsFrame";
+import EventsFrame from "@/components/events-page/EventsFrame";
 
 export default async function GroupsPage() {
 	// if user is not authenticated, redirect them to login
