+++ frontend/app/groups/page.tsx
@@ -1,8 +1,8 @@
 import { getCurrentSession } from "@/actions/users.server";
-import PageLayout from "@/components/PageLayout";
+import PageLayout from "@/components/shared/PageLayout";
 import { UserDisplayData } from "@/schemas/fe.schema";
 import { redirect } from "next/navigation";
-import GroupsFrame from "@/components/GroupsFrame";
+import GroupsFrame from "@/components/groups-page/GroupsFrame";
 
 export default async function GroupsPage() {
 	// if user is not authenticated, redirect them to login
@@ -13,7 +13,7 @@ export default async function GroupsPage() {
 
 	return (
 		<PageLayout>
-			 <GroupsFrame GroupsData={userData.groups} />
+			<GroupsFrame GroupsData={userData.groups} />
 		</PageLayout>
 	);
 }
