+++ frontend/app/page.tsx
@@ -1,6 +1,6 @@
 import { getCurrentSession } from "@/actions/users.server";
-import HomeFrame from "@/components/HomeFrame";
-import PageLayout from "@/components/PageLayout";
+import HomeFrame from "@/components/home-page/HomeFrame";
+import PageLayout from "@/components/shared/PageLayout";
 import { UserDisplayData } from "@/schemas/fe.schema";
 import { redirect } from "next/navigation";
 
@@ -13,9 +13,9 @@ export default async function Home() {
 
 	// Safely log groups and events if they exist
 	if (userData.groups && userData.groups.length > 0) {
-		console.log('User has groups:', userData.groups);
+		console.log("User has groups:", userData.groups);
 	} else {
-		console.log('User has no groups yet');
+		console.log("User has no groups yet");
 	}
 	return (
 		<PageLayout>
