+++ frontend/components/groups-page/GroupCard.tsx
