+++ frontend/components/users/ProfileFrame.tsx
@@ -1,6 +1,6 @@
 import { UserDisplayData } from "@/schemas/fe.schema";
 import Image from "next/image";
-import Button from "@/components/Button";
+import Button from "@/components/shared/Button";
 import Link from "next/link";
 
 export default function ProfileFrame({
